fix: extracted crit value cell into separate component
