+++ src/components/styles/CharactersTable.css
@@ -30,17 +30,6 @@
     background-color: var(--mantine-primary-color-filled);
     color: black;
 }
-.crit-cell {
-    display: flex;
-    position: relative;
-    gap: 24px;
-    white-space: nowrap;
-    /* font-family: "JetBrains Mono", monospace; */
-}
-.crit-cell > div:first-child {
-    text-align: end;
-    width: 72px;
-}
 
 .character-card-row > td {
     background-color: var(--table-striped-color);
